% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{adj_R2}
\alias{adj_R2}
\title{Calculates the adjusted r-squared statistic between predicted and ground truth continuous variables.}
\usage{
adj_R2(pred, truth, n_p)
}
\arguments{
\item{pred}{A vector of predicted values}

\item{truth}{A vector of true values}

\item{n_p}{The number of model parameters, excluding the intercept}
}
\value{
The adjusted r-squared statistic between predicted and ground truth continuous variables.
}
\description{
This function accepts three arguments: the predicted values `pred` and the true values `truth` of a
continuous variable, and the number of model parameters `n_p`, and calculates the adjusted r-squared statistic.
}
\examples{
pred <- c(1.2, 2.4, 3.6, 4.8, 6.0)
truth <- c(1.0, 2.5, 3.2, 5.0, 5.5)
n_p <- 2
adj_R2(pred, truth, n_p)

# Output:
# [1] 0.8672337

}
