% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{classes}
\alias{classes}
\title{Returns a character vector containing the classes of each variable in a tibble x.}
\usage{
classes(x)
}
\arguments{
\item{x}{A tibble}
}
\value{
A character vector containing the classes of each variable in `x`.
}
\description{
This function accepts a tibble `x` and returns a character vector containing the classes of each
variable in `x`. It is similar to the `names` function, but instead of returning the variable names,
it returns the variable classes.
}
\examples{
# Load the tibble package
library(tibble)

# Create a tibble
x <- tibble(a = 1:5, b = c("foot", "bar", "bazar", "math", "queen"), c = as.Date("2023-01-01") + 0:4)

# Get the classes of each variable
classes(x)

# Output:
#   a        b        c
# "integer" "character" "Date"

}
