% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{df_scale}
\alias{df_scale}
\title{Returns a tibble x in which the numeric variables have been scaled with scale.}
\usage{
df_scale(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{A tibble}

\item{center}{Logical indicating if the variables should be centered (default is TRUE).}

\item{scale}{Logical indicating if the variables should be scaled (default is TRUE).}
}
\value{
A new tibble in which the numeric variables have been scaled with `scale`.
}
\description{
This function accepts a tibble `x` and returns a new tibble in which the numeric variables have
been scaled with the `scale` function. The `center` and `scale` arguments can be used to control
the scaling behavior.
}
\examples{
# Load the tibble package
library(tibble)

# Create a tibble
x <- tibble(a = c(1, 2, 3), b = c(4, 5, 6), c = c(7, 8, 9))

# Scale the numeric variables
df_scale(x)

# Output:
# A tibble: 3 × 3
      a     b     c
  <dbl> <dbl> <dbl>
1    -1    -1    -1
2     0     0     0
3     1     1     1

}
