% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{f1}
\alias{f1}
\title{Calculate the F1 score of a binary classification model}
\usage{
f1(pred, truth)
}
\arguments{
\item{pred}{A vector of predicted labels (0 or 1)}

\item{truth}{A vector of true labels (0 or 1)}
}
\value{
The F1 score of the binary classification model predictions.
}
\description{
This function takes two vectors, the predicted labels `pred` (0 or 1) and the true
labels `truth` (0 or 1) of a binary classification model, and calculates the F1 score,
which is the harmonic mean of precision and recall.
}
\examples{
pred <- c(0, 1, 1, 0, 1)
truth <- c(0, 1, 1, 1, 0)
f1(pred, truth)

}
