% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{is_max}
\alias{is_max}
\title{is_max: Returns a logical vector indicating where the maximum value of an atomic vector is found}
\usage{
is_max(x, na.rm = T)
}
\arguments{
\item{x}{An atomic vector to be checked for its maximum value}

\item{na.rm}{A logical value indicating whether to remove NA values before finding the maximum.}
}
\value{
A logical vector with TRUE values indicating where the maximum value of "x" is found.
}
\description{
This function takes an atomic vector "x" and returns a logical vector with TRUE values indicating the locations
where the maximum value of "x" is found. The function also includes an optional parameter "na.rm" which, if set
to TRUE, removes any NA values from the input vector before finding the maximum value.
}
\examples{
is_max(c(11, 21, 31, 22, 42, 11))
is_max(c(NA, 2, 3, NA, 4, NA), na.rm = TRUE)

}
