% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{is_min}
\alias{is_min}
\title{is_min: Returns a logical vector indicating where the minimum value of an atomic vector is found}
\usage{
is_min(x, na.rm = T)
}
\arguments{
\item{x}{An atomic vector to be checked for its minimum value}

\item{na.rm}{A logical value indicating whether to remove NA values before finding the minimum. Default is TRUE.}
}
\value{
A logical vector with TRUE values indicating where the minimum value of "x" is found.
}
\description{
This function takes an atomic vector "x" and returns a logical vector with TRUE values indicating the
locations where the minimum value of "x" is found. The function also includes an optional parameter
"na.rm" which, if set to TRUE, removes any NA values from the input vector before finding the minimum value.
}
\examples{
is_min(c(1, 2, 3, 1, 4, 1))
is_min(c(NA, 2, 3, NA, 4, NA), na.rm = TRUE)
}
