% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_f}
\alias{log_likelihood_f}
\title{Returns the log-likelihood of a sample x under the F distribution.}
\usage{
log_likelihood_f(x, df1, df2)
}
\arguments{
\item{x}{A numeric vector}

\item{df1}{Degrees of freedom for the numerator of the F distribution}

\item{df2}{Degrees of freedom for the denominator of the F distribution}
}
\value{
The log-likelihood of `x` under the F distribution with `df1` and `df2` degrees of freedom.
}
\description{
This function accepts a numeric vector `x` and degrees of freedom `df1` and `df2`, and returns the
log-likelihood of `x` under the F distribution with `df1` and `df2` degrees of freedom.
}
\examples{
log_likelihood_f(c(1,2,3), 2, 5)

}
