% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_norm}
\alias{log_likelihood_norm}
\title{Returns the log-likelihood of a sample x under the normal distribution.}
\usage{
log_likelihood_norm(x, mean, sd)
}
\arguments{
\item{x}{A numeric vector}

\item{mean}{The mean of the normal distribution}

\item{sd}{The standard deviation of the normal distribution}
}
\value{
The log-likelihood of the sample `x` under the normal distribution with mean `mean` and
standard deviation `sd`.
}
\description{
This function accepts a numeric vector `x`, and parameters `mean` and `sd` of a normal distribution,
and returns the log-likelihood of observing the sample `x` under the normal distribution
with mean `mean` and standard deviation `sd`.
}
\examples{
# Compute the log-likelihood of a sample
x <- c(1.2, 2.1, 3.4, 4.3, 5.6)
log_likelihood_norm(x, 3, 1.5)

# Output:
# [1] -8.652607

}
