% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_t}
\alias{log_likelihood_t}
\title{Returns the log-likelihood of a sample x under the t distribution.}
\usage{
log_likelihood_t(x, df)
}
\arguments{
\item{x}{A numeric vector}

\item{df}{Degrees of freedom for the t distribution}
}
\value{
The log-likelihood of `x` under the t distribution with `df` degrees of freedom.
}
\description{
This function accepts a numeric vector `x` and a degrees of freedom `df`, and returns the log-likelihood
of `x` under the t distribution with `df` degrees of freedom.
}
\examples{
log_likelihood_t(c(1,2,3), 4)

}
