% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_unif}
\alias{log_likelihood_unif}
\title{Returns the log-likelihood of a sample x under the uniform distribution.}
\usage{
log_likelihood_unif(x, min, max)
}
\arguments{
\item{x}{A numeric vector of observations.}

\item{min}{The minimum value of the uniform distribution.}

\item{max}{The maximum value of the uniform distribution.}
}
\value{
A scalar value representing the log-likelihood of the sample `x` under the uniform distribution.
}
\description{
This function calculates the log-likelihood of a sample `x` under the uniform distribution,
where the minimumand maximum values are specified by `min` and `max`, respectively.
}
\examples{
# Calculate the log-likelihood of a sample under the uniform distribution
x <- c(0.2, 0.3, 0.4, 0.5)
log_likelihood_unif(x, 0, 1)
# Output:
# [1] -5.545177

}
