% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rando}
\alias{rando}
\title{Rando: A wrapper around sample that tests whether x is an atomic vector or dataframe-like object
and then returns either n samples or n rows as appropriate.}
\usage{
rando(x, n = 1, replace = T)
}
\arguments{
\item{x}{An atomic vector or data frame-like object}

\item{n}{The number of samples or rows to return}

\item{replace}{Logical; should sampling be with replacement}
}
\value{
A vector or data frame-like object containing the sampled elements or rows.
}
\description{
This function checks whether the input `x` is an atomic vector or a
data frame-like object (data frame, matrix, or list) and returns either
`n` samples or `n` rows accordingly. If the input does not match any of these
types, an error message will be shown.
}
\examples{
# sample from an atomic vector
rando(15:25, 3, replace = TRUE)
# sample from a dataframe
df = data.frame(x = 10:50, y = 60:100)
rando(df, 2, replace = TRUE)
}
