% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rep_mat}
\alias{rep_mat}
\title{rep_mat: Replicates a matrix or dataframe by repeating its rows or columns M or N times}
\usage{
rep_mat(x, M = 1, N = 1)
}
\arguments{
\item{x}{A matrix or dataframe to be replicated}

\item{M}{An integer specifying the number of times to replicate the rows of "x"}

\item{N}{An integer specifying the number of times to replicate the columns of "x"}
}
\value{
A new matrix created by replicating the rows or columns of "x" M or N times.
}
\description{
This function takes a matrix or dataframe "x" and replicates its rows or columns M or N times,
returning a new matrix. The values of M and N determine the number of times the rows or columns of "x"
are repeated, respectively.
If M and N are both set to 1, the function returns the input matrix or dataframe as is.
}
\examples{
x <- matrix(c(1, 2, 3, 4), nrow = 2, ncol = 2)
rep_mat(x, M = 2, N = 3)
}
